+++ src/quick-lint-js/diag/diagnostic-metadata-generated.h
@@ -334,7 +334,6 @@ namespace quick_lint_js {
   QLJS_DIAG_TYPE_NAME(Diag_Unclosed_String_Literal) \
   QLJS_DIAG_TYPE_NAME(Diag_Unclosed_JSX_String_Literal) \
   QLJS_DIAG_TYPE_NAME(Diag_Unclosed_Template) \
-  QLJS_DIAG_TYPE_NAME(Diag_Unexpected_At_Character) \
   QLJS_DIAG_TYPE_NAME(Diag_Unexpected_Arrow_After_Expression) \
   QLJS_DIAG_TYPE_NAME(Diag_Unexpected_Arrow_After_Literal) \
   QLJS_DIAG_TYPE_NAME(Diag_Unexpected_Backslash_In_Identifier) \
@@ -409,7 +408,7 @@ namespace quick_lint_js {
   /* END */
 // clang-format on
 
-inline constexpr int Diag_Type_Count = 395;
+inline constexpr int Diag_Type_Count = 394;
 
 extern const Diagnostic_Info all_diagnostic_infos[Diag_Type_Count];
 }