+++ docs/CHANGELOG.md
@@ -6,6 +6,12 @@ based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 quick-lint-js' version numbers are arbitrary. quick-lint-js does *not* adhere to
 Semantic Versioning.
 
+## Unreleased
+
+### Added
+
+* quick-lint-js now understands decorators.
+
 ## 2.16.0 (2023-09-06)
 
 [Downloads](https://c.quick-lint-js.com/releases/2.16.0/)